requirements.txt
cycler==0.10.0
glfw==1.11.0
kiwisolver==1.2.0
matplotlib==3.2.1
numpy==1.18.4
PyOpenGL==3.1.5
PyOpenGL-accelerate==3.1.5
pyparsing==2.4.7
python-dateutil==2.8.1
python-interface==1.5.5
six==1.14.0
